# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Fininzen API",
        "version": "0.13.0",
        "description": "Wealth-management API (expenses + portfolio)."
    },
    "paths": {
...
        "/api/expenses/description-suggestions/": {
            "get": {
                "operationId": "expenses_description_suggestions_retrieve",
                "description": "GET /api/expenses/description-suggestions/?category_id=<id>&q=<prefix>",
                "tags": [
                    "expenses"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/import-csv/": {
            "post": {
                "operationId": "expenses_import_csv_create",
                "description": "POST /api/expenses/import-csv/\nBody: { rows: [{date, description, amount, category_name?}] }\nCrea le spese corrispondenti. Match categoria per nome (case-insensitive).",
                "tags": [
                    "expenses"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/last-account/": {
            "get": {
                "operationId": "expenses_last_account_retrieve",
                "description": "GET /api/expenses/last-account/?category=<id>\n\nReturns {\"linked_asset\": <id|null>}: the account (linked_asset) of the\nmost recent expense the user filed under that category. Powers the\ncash-flow \"autofill last account on category\" preference.",
                "tags": [
                    "expenses"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/monthly/": {
            "get": {
                "operationId": "expenses_monthly_retrieve",
                "description": "CRUD completo per le spese con filtri per periodo.",
                "tags": [
                    "expenses"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/quick-add/": {
            "post": {
                "operationId": "expenses_quick_add_create",
                "description": "POST /api/expenses/quick-add/ — Automation-only expense creation\n(iOS Shortcuts / Apple Pay NFC). Requires an API token with\nexpenses:write scope; JWT/session auth is never even considered for\nthis action (authentication_classes is overridden above), so a\nbrowser session cannot reach it. Category is resolved by name\n(case-insensitive); an unmatched or omitted name falls back to a\nper-user \"Da categorizzare\" category rather than failing the request.",
                "tags": [
                    "expenses"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "apiTokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
        "/api/expenses/reset/": {
            "post": {
                "operationId": "expenses_reset_create",
                "description": "POST /api/expenses/reset/ — Cancella tutte le spese.\n\nBody required: {\"confirm\": true}. Server-side guard against accidental\nor CSRF-style invocation (UI modal alone is not enough).",
                "tags": [
                    "expenses"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/seed-demo/": {
            "post": {
                "operationId": "expenses_seed_demo_create",
                "description": "POST /api/expenses/seed-demo/\nCancella spese + asset esistenti e genera dati demo casuali:\n- 6 mesi di spese (8-15/mese) e entrate (1-3/mese)\n- 4-7 asset con valori realistici\nCrea categorie e tipi di investimento di default se non esistono.",
                "tags": [
                    "expenses"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/summary/": {
            "get": {
                "operationId": "expenses_summary_retrieve",
                "description": "CRUD completo per le spese con filtri per periodo.",
                "tags": [
                    "expenses"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/expenses/trends/": {
            "get": {
                "operationId": "expenses_trends_retrieve",
                "description": "Daily income and expense totals for the trailing ten years.",
                "tags": [
                    "expenses"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 7
                    },
                    "icon": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "category_type": {
                        "$ref": "#/components/schemas/CategoryTypeEnum"
                    },
                    "parent": {
                        "type": "integer",
                        "nullable": true
                    },
                    "expense_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "subcategory_expense_count": {
                        "type": "string",
                        "readOnly": true
                    },
                    "subcategories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subcategory"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "expense_count",
                    "id",
                    "name",
                    "subcategories",
                    "subcategory_expense_count"
                ]
            },
...
            "CategoryTypeEnum": {
                "enum": [
                    "expense",
                    "income"
                ],
                "type": "string",
                "description": "* `expense` - Uscita\n* `income` - Entrata"
            },
...
            "Expense": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "amount": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "category": {
                        "type": "integer"
                    },
                    "category_detail": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Category"
                            }
                        ],
                        "readOnly": true
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "linked_asset": {
                        "type": "integer",
                        "nullable": true
                    },
                    "linked_asset_name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "is_verified": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "amount",
                    "category",
                    "category_detail",
                    "created_at",
                    "date",
                    "description",
                    "id",
                    "linked_asset_name"
                ]
            },
            "ExpenseRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "amount": {
                        "type": "string",
                        "format": "decimal",
                        "pattern": "^-?\\d{0,8}(?:\\.\\d{0,2})?$"
                    },
                    "category": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "linked_asset": {
                        "type": "integer",
                        "nullable": true
                    },
                    "is_verified": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "amount",
                    "category",
                    "date",
                    "description"
                ]
            },
...
            "Subcategory": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 7
                    },
                    "icon": {
                        "type": "string",
                        "maxLength": 10
                    },
                    "category_type": {
                        "$ref": "#/components/schemas/CategoryTypeEnum"
                    },
                    "parent": {
                        "type": "integer",
                        "nullable": true
                    },
                    "expense_count": {
                        "type": "integer",
                        "readOnly": true
                    }
                },
                "required": [
                    "expense_count",
                    "id",
                    "name"
                ]
            },
...
        },
        "securitySchemes": {
            "apiTokenAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Opaque"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}